requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2
email-validator==2.2.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2
